change float to byte

+++ ProjektJA1/C/dllmain.cpp
@@ -1,15 +1,16 @@
 ﻿// dllmain.cpp : Defines the entry point for the DLL application.
 #include "pch.h"
+#include <stdint.h>
 
-void GrayScale(float* pixels, int beg, int end)
+void GrayScale(uint8_t* pixels, int beg, int end)
 {
     for (int i = beg; i < end; i += 4)
     {
-        float b = pixels[i];
-        float g = pixels[i + 1];
-        float r = pixels[i + 2];
+        uint8_t b = pixels[i];
+        uint8_t g = pixels[i + 1];
+        uint8_t r = pixels[i + 2];
 
-        float avg = (r + g + b) / 3;
+        uint8_t avg = (r + g + b) / 3;
 
         pixels[i] = avg;
         pixels[i + 1] = avg;
@@ -18,7 +19,7 @@ void GrayScale(float* pixels, int beg, int end)
 }
 
 
-extern "C" __declspec(dllexport) void GreyScaleCFunc(float* pixels, int beg, int end)
+extern "C" __declspec(dllexport) void GreyScaleCFunc(uint8_t * pixels, int beg, int end)
 {
     GrayScale(pixels, beg, end);
 }
