+++ ProjektJA1/ProjektJA1/Mechanism/GrayScaleC.cs
@@ -6,9 +6,9 @@ namespace ProjektJA.Mechanism
     internal class GrayScaleC : GrayScaleInterface
     {
         [DllImport(@"D:\studia\github\JAproj\ProjektJA1\x64\Debug\C.dll")]
-        static extern void GreyScaleCFunc(float[] pixels, int beg, int end);
+        static extern void GreyScaleCFunc(byte[] pixels, int beg, int end);
 
-        public override void ExecuteEffect(float[] pixels, int beg, int end)
+        public override void ExecuteEffect(byte[] pixels, int beg, int end)
         {
             GreyScaleCFunc(pixels, beg, end);
         }
