+++ ProjektJA1/ProjektJA1/Mechanism/GrayScaleInterface.cs
@@ -4,6 +4,6 @@ namespace ProjektJA.Mechanism
 {
     public abstract class GrayScaleInterface
     {
-        public abstract void ExecuteEffect(float[] pixels, int beg, int end); 
+        public abstract void ExecuteEffect(byte[] pixels, int beg, int end); 
     }
 }
